ai-service/requirements.txt
requests~=2.32.4
beautifulsoup4~=4.13.4
scikit-learn~=1.7.0
sentence-transformers
nltk~=3.9.1
google-generativeai
python-dotenv
numpy~=2.3.1
Flask~=3.1.1